Shrink: int parameter for passed()
All callers pass int literals. bipush is cheaper than ldc2_w
for the double version. -2 bytes (1458 → 1456).

+++ src/pez/mini/Pugilist.java
@@ -235,7 +235,7 @@ public class Pugilist extends AdvancedRobot {
             registerVisits(fastFactors, 1);
         }
 
-        public boolean passed(double distanceOffset) {
+        public boolean passed(int distanceOffset) {
             return distanceFromGun > gunLocation.distance(targetLocation) + distanceOffset;
         }
